Motor torque in nm (#91)
* Ramp management improved for torque and speed control

--> Management of motor ramp using slope parameter instead of duration

* Changed stuff based on Francois's PR review

* Added foldback to motor control layer

Before, foldback were managed by powertrain module, so vehicle layer. A better and simple control can be achieved by letting motor control layer manage some foldback such as speed and temperature

* Fixed indentation of mc_tasks and md_interface

* Bug fix found during load testing

* Added signal filtering files

* LP FO Butterworth and recursive average working good

* Application specific tuning to change default values

* Added ASSERT

* Torque control in cNm (Nm/100)

Also added formula to set current/torque ratio from motor characteristics

* Forgot to change some constant to take into account Nm/100 unit

+++ src/GNR2_Firmware/revA/grn2_ra6t2/FTEX/inc/middleware_vehicle/throttle.h
@@ -23,16 +23,16 @@ typedef struct
 	uint16_t hLowPassFilterBW2;
 	
 	uint16_t hOffsetThrottle;            // Offset of ADC value vs throttle
-	uint8_t bSlopeThrottle;              // Gain factor of ADC value vs throttle
-	uint8_t bDivisorThrottle;            // Scaling factor of ADC value vs throttle   
+	uint16_t bSlopeThrottle;              // Gain factor of ADC value vs throttle
+	uint16_t bDivisorThrottle;            // Scaling factor of ADC value vs throttle   
 	
 	uint16_t hOffsetTorque;          	 // Offset of throttle vs torque 
-	int8_t bSlopeTorque;                 // Gain factor of throttle vs torque   
-	uint8_t bDivisorTorque;              // Scaling factor of throttle vs torque   
+	int16_t bSlopeTorque;                 // Gain factor of throttle vs torque   
+	uint16_t bDivisorTorque;              // Scaling factor of throttle vs torque   
 	
 	uint16_t hOffsetSpeed;               // Offset of throttle vs speed 
-	int8_t bSlopeSpeed;                  // Gain factor of throttle vs speed   
-	uint8_t bDivisorSpeed;               // Scaling factor of throttle vs speed   
+	int16_t bSlopeSpeed;                  // Gain factor of throttle vs speed   
+	uint16_t bDivisorSpeed;               // Scaling factor of throttle vs speed   
 	
 	uint16_t hDetectionThreshold;        // Minimum throttle at which throttle is detected
